+++ scourgify/__init__.py
@@ -8,5 +8,6 @@ All rights reserved
 # Local Imports
 from scourgify.normalize import (
     get_geocoder_normalized_addr,
-    normalize_address_record
+    normalize_address_record,
+    NormalizeAddress
 )